LayoutTests/editing/deleting/delete-br-in-last-table-cell-live-range-expected.txt

| <html>
|   <!--  webkit-test-runner [ LiveRangeSelectionEnabled=true ]  -->
|   <head>
|     "\n"
|     <script>
|       language="JavaScript"
|       src="../../resources/dump-as-markup.js"
|       type="text/JavaScript"
|     "\n"
|     <script>
|       language="JavaScript"
|       src="../editing.js"
|       type="text/JavaScript"
|     "\n"
|     <script>
|       "\nfunction runTest()\n{\n    br = document.getElementById("lastBR");\n    sel = window.getSelection();\n    sel.setBaseAndExtent(br, 0, br, 0);\n    deleteCommand();\n}\n"
|     "\n"
|   "\n\n"
|   <body>
|     "\n"
|     <p>
|       <a>
|         href="https://bugs.webkit.org/show_bug.cgi?id=35369"
|         "Bug 35369"
|       " and "
|       <a>
|         href="https://bugs.webkit.org/show_bug.cgi?id=35632"
|         "Bug 35632"
|     "\n"
|     <p>
|       "Executing a delete command when positioned before a BR in a table cell (esp. the last)"
|     "\n"
|     <div>
|       contenteditable=""
|       "\n"
|       <table>
|         id="table"
|         <tbody>
|           <tr>
|             <td>
|               "1"
|             <td>
|               "2"
|           <tr>
|             <td>
|               "3"
|             <td>
|               "4<#selection-caret>"
|               <br>
|                 id="lastBR"
|       "\n"
|     "\n"
|     <script>
|       "\nrunTest();\n"
|     "\n\n\n"
